Improve test of compressors config parser
